:art: refactor: 유틸리티 클래스들을 한 폴더로 위치 변경
