+++ src/main/java/racingcar/Validation/TurnCountValidator.java
@@ -1,18 +1,14 @@
 package racingcar.Validation;
 
 import racingcar.Enum.ErrorMessage;
+import racingcar.Utilities.NumberParser;
 
 public class TurnCountValidator {
-    public static int validateTurnCount(String turnCount){
-        try {
-            Integer intTurnCount = Integer.parseInt(turnCount);
-            if(intTurnCount < 1){
-                throw new IllegalArgumentException(ErrorMessage.INVALID_TURN_NUMBER.getMessage());
-            }
-            return intTurnCount;
-
-        } catch (NumberFormatException e) {
-            throw new IllegalArgumentException(ErrorMessage.INVALID_TURN_NUMBER.getMessage());
+    public static int validateTurnCount(String turnCount) {
+        int intTurnCount = NumberParser.parseToInt(turnCount);
+        if (intTurnCount < 1) {
+            throw new IllegalArgumentException(ErrorMessage.INVALID_INT_NUMBER.getMessage());
         }
+        return intTurnCount;
     }
 }
